+++ R/package-init.R
@@ -1,8 +1,13 @@
+#' @keywords internal
+"_PACKAGE"
+
+
 #' @import dplyr
 #' @importFrom stats dpois quantile rpois dbinom weighted.mean
 #' @importFrom reldist wtd.quantile
 #' @importFrom splines bs
 #' @importFrom tibble is_tibble
+#' @importFrom vetiver vetiver_create_description vetiver_create_ptype handler_startup handler_predict vetiver_create_meta
 NULL
 
 #' Sample Patient Characteristics Inputs